interface works, just need to implement QuadDrone.cs and fix any scene issues moving files around caused

+++ Assets/Scripts/DroneInterface/DroneCommands.cs
@@ -3,10 +3,10 @@ namespace DroneInterface
     interface IDroneCommands
     {
         // The drone will navigate to the 3D point (local_north, local_east, local_down)
-        void Goto(float latitude, float longitude, float altitude);
+        void Goto(double latitude, double longitude, double altitude);
 
         // The drone will fly up or down to the desired relative altitude
-        void Hover(float altitude);
+        void Hover(double altitude);
 
         // Arm/disarm the drone
         void Arm(bool arm);
@@ -15,17 +15,17 @@ namespace DroneInterface
         void TakeControl(bool arm);
 
         // The drone will fly at the commanded orientation (roll, pitch, yaw) and a vertical velocity
-        void SetAttitude(float roll, float pitch, float yaw, float velocity);
+        void SetAttitude(double roll, double pitch, double yaw, double velocity);
 
         // The drone will fly with the following angular rates (roll_rate, pitch_rate, yaw_rate) and collective thrust (m/s^2)
-        void SetAttitudeRate(float rollRate, float pitchRate, float yawRate, float thrust);
+        void SetAttitudeRate(double rollRate, double pitchRate, double yawRate, double thrust);
 
-        // The drone will fly at the commanded velocity (north_velocity, east_velocity, vertical_velocity) and heading
-        void SetVelocity(float northVelocity, float eastVelocity, float verticalVelocity, float heading);
+        // The drone will fly at the commanded velocity (north velocity, east velocity, vertical velocity) and heading
+        void SetVelocity(double northVelocity, double eastVelocity, double verticalVelocity, double heading);
 
         // Command the following thrust (possible RPM) to the motors directly
         // TODO: set appropriate variable names
-        void SetMotors(float a, float b, float c, float d);
+        void SetMotors(double a, double b, double c, double d);
         // ...
     }
 }
